Add the option to define a channel time offset in the sources.xml
This allows epgimport to correct start- and endtimes of events, as several xmltv sources use local time with +0000 instead of the correct timezone offset. To use this, define the offset like so:
````
<source type="gen_xmltv" nocheck="1" offset="-0200" channels="/etc/epgimport/some.xmltvfile.channels.xml">
 <description>Some XMLTV file</description>
 <url>http://epg.example.org/some,xlmtvfile.gz></url>
</source>
````
in this case because the times in the file are in GMT+2, but the offset is +0000, indicating GMT was used. The offset "-0200" corrects GMT+2 to GMT.