[Fixed] Not working import due to broken rating and category/genre
****** WARNING ******
NEVER EVER change keys of category dict!
XMLTV generators rely on that to specify correct genre so to be imported.
If you need you can add new entries but NEVER remove or change the keys.

***** Language *****
Currently enigma2 doesnt support specification of language during import. Only rating can be imported. So you cant pass language to the import procedure like that!

@Lululla @Belfagor2005
When you make changes you have to be aware for how enigma works. DONT rely on AI to do your job because this is what happens!!!!